Ensure the x64 cudadevrt.lib gets packaged (Fixes #12)
There are two versions of cudadevrt.lib in the toolkit installer - an
x64 version and a Win32 version. The Win32 version was getting picked up
by the extractor, but the x64 version is required. This commit ensures
that x64 is in the path of cudadevrt.lib before copying, so that the
correct version is packaged.